Refactor package layout
Moved Email services to their own package
