+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/coheventmapper/actions/DecisionIssuedEventAction.java
@@ -3,8 +3,8 @@ package uk.gov.hmcts.reform.sscscorbackend.coheventmapper.actions;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
-import uk.gov.hmcts.reform.sscscorbackend.service.CorEmailService;
-import uk.gov.hmcts.reform.sscscorbackend.service.EmailMessageBuilder;
+import uk.gov.hmcts.reform.sscscorbackend.service.email.CorEmailService;
+import uk.gov.hmcts.reform.sscscorbackend.service.email.EmailMessageBuilder;
 import uk.gov.hmcts.reform.sscscorbackend.service.StoreOnlineHearingTribunalsViewService;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.CohEventActionContext;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.PdfData;
