+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/service/email/DecisionEmailService.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.sscscorbackend.service;
+package uk.gov.hmcts.reform.sscscorbackend.service.email;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
 import uk.gov.hmcts.reform.sscscorbackend.domain.TribunalViewResponse;
+import uk.gov.hmcts.reform.sscscorbackend.service.JuiUrlGenerator;
 
 @Service
 public class DecisionEmailService {
