+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/service/email/AppellantStatementEmailServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscscorbackend.service;
+package uk.gov.hmcts.reform.sscscorbackend.service.email;
 
 import static org.mockito.Mockito.*;
